move currentEventResourceAtom to core/shared_state
