+++ src/core/shared_state/currentEventResource.ts
