+++ src/features/current_event/atoms/currentEventGeometry.ts
@@ -1,6 +1,6 @@
 import { createAtom } from '~utils/atoms';
 import { focusedGeometryAtom } from '~core/shared_state';
-import { currentEventResourceAtom } from './currentEventResource';
+import { currentEventResourceAtom } from '../../../core/shared_state/currentEventResource';
 import type { EventWithGeometry } from '~core/types';
 
 export const currentEventGeometryAtom = createAtom(
