+++ src/features/events_list/components/CurrentEvent/CurrentEvent.tsx
@@ -2,12 +2,12 @@ import { useAtom } from '@reatom/react';
 import { ErrorMessage } from '~components/ErrorMessage/ErrorMessage';
 import { LoadingSpinner } from '~components/LoadingSpinner/LoadingSpinner';
 import { i18n } from '~core/localization';
-import { currentEventResourceAtom } from '~features/current_event/atoms/currentEventResource';
+import { currentEventResourceAtom } from '~core/shared_state/currentEventResource';
 import { createStateMap } from '~utils/atoms';
 import { EpisodeTimelineToggle } from '../EpisodeTimelineToggle/EpisodeTimelineToggle';
 import { EventCard } from '../EventCard/EventCard';
 
-export function CurrentEventFromResource({ hasTimeline }: { hasTimeline?: boolean }) {
+export function CurrentEvent({ hasTimeline }: { hasTimeline?: boolean }) {
   const [{ data, error, loading }] = useAtom(currentEventResourceAtom);
 
   const statesToComponents = createStateMap({
